Solved Clients, Bookings, Rooms.

+++ src/main/java/com/proiectip/boat/owners/OwnerController.java
@@ -51,28 +51,39 @@ public class OwnerController {
 
     // adauga id-ul proprietatii in lista de proprietati a ownerului
     @PutMapping("/addPropertyInOwner")
-    public ResponseEntity<String> addPropertyInOwner(@RequestBody Properties property){
-        String idOwner = property.getOwner();
-        Optional<Owners> owner = ownersRepository.findById(idOwner);
-        if(owner.isPresent()) {
-            propertiesRepository.save(property);
-            owner.get().getProperties().add(property.getId());
-            ownersRepository.save(owner.get());
-            return new ResponseEntity("Property added succesfully!", HttpStatus.OK);
+    public ResponseEntity<String> addPropertyInOwner(@RequestBody Map<String,String> map){
+        String idOwner = map.get("owner");
+        Owners owner = ownersRepository.findById(idOwner).get();
+        if(owner == null) {
+            return new ResponseEntity("Owner not found!", HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity("Owner not found!", HttpStatus.BAD_REQUEST);
+        if(!owner.isAccepted())
+            return new ResponseEntity("Owner not accepted!", HttpStatus.BAD_REQUEST);
+        String name = map.get("name");
+        String location = map.get("location");
+        String description = map.get("description");
+        String typeOfProperty = map.get("typeOfProperty");
+        Properties property = new Properties(owner, name, location, description, typeOfProperty);
+
+        propertiesRepository.save(property);
+        owner.getProperties().add(property.getId());
+        ownersRepository.save(owner);
+        return new ResponseEntity("Property added succesfully!", HttpStatus.OK);
     }
 
     @DeleteMapping("/deletePropertyInOwner")
     public ResponseEntity<String> deletePropertyInOwner(@RequestBody Map<String, String> map){
-        String idOwner = map.get("owner");
         String idProperty = map.get("_id");
-        Optional<Owners> owner = ownersRepository.findById(idOwner);
+        Properties property = propertiesRepository.findById(idProperty).get();
+        if(property == null) {
+            return new ResponseEntity("Property not found!", HttpStatus.BAD_REQUEST);
+        }
+        Optional<Owners> owner = ownersRepository.findById(property.getOwner().getId());
         if(owner.isPresent()) {
             owner.get().getProperties().remove(idProperty);
             ownersRepository.save(owner.get());
             propertiesRepository.deleteById(idProperty);
-            return new ResponseEntity("Property deleted succesfully!", HttpStatus.OK);
+            return new ResponseEntity("Property deleted successfully!", HttpStatus.OK);
         }
         return new ResponseEntity("Owner not found!", HttpStatus.BAD_REQUEST);
     }
